requirements.txt
Flask
numpy
pandas
Pillow
python-dateutil
python-dotenv
torch
torchvision
pyTelegramBotAPI
